requirements.txt
numpy
pydub
scipy
matplotlib
ffmpeg-python
